requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
bleach==6.0.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.4
django-js-asset==2.1.0
django-mptt==0.14.0
django-ranged-response==0.2.0
django-recaptcha==3.0.0
django-simple-captcha==0.5.18
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
kombu==5.3.1
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2023.3
redis==5.0.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
